Split specific and generic rector
