+++ packages/CodingStyle/tests/Rector/FuncCall/FunctionCallToConstantRector/Fixture/fixture.php.inc
@@ -2,11 +2,13 @@
 
 namespace Rector\CodingStyle\Tests\Rector\FuncCall\FunctionCallToConstantRector\Fixture;
 
-class PiClass
+class SomeClass
 {
     public function run()
     {
         $value = pi();
+
+        $value2 = php_sapi_name();
     }
 }
 
@@ -16,11 +18,13 @@ class PiClass
 
 namespace Rector\CodingStyle\Tests\Rector\FuncCall\FunctionCallToConstantRector\Fixture;
 
-class PiClass
+class SomeClass
 {
     public function run()
     {
         $value = M_PI;
+
+        $value2 = PHP_SAPI;
     }
 }
