--- packages/CodingStyle/tests/Rector/FuncCall/FunctionCallToConstantRector/Fixture/sapi-name.php.inc
@@ -1,27 +1,0 @@
-<?php
-
-namespace Rector\CodingStyle\Tests\Rector\FuncCall\FunctionCallToConstantRector\Fixture;
-
-class SapiNameClass
-{
-    public function run()
-    {
-        $value = php_sapi_name();
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\CodingStyle\Tests\Rector\FuncCall\FunctionCallToConstantRector\Fixture;
-
-class SapiNameClass
-{
-    public function run()
-    {
-        $value = PHP_SAPI;
-    }
-}
-
-?>
