+++ packages/CodingStyle/tests/Rector/FuncCall/VersionCompareFuncCallToConstantRector/Fixture/skip-version-compare.php.inc
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rector\CodingStyle\Tests\Rector\FuncCall\VersionCompareFuncCallToConstantRector\Fixture;
+
+class SkipVersionCompareClass
+{
+    public function run()
+    {
+        version_compare($this->version, '5.3.0', '<');
+    }
+}
+
+?>
