+++ packages/CodingStyle/tests/Rector/FuncCall/VersionCompareFuncCallToConstantRector/Fixture/version-compare.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\CodingStyle\Tests\Rector\FuncCall\FunctionCallToConstantRector\Fixture;
+namespace Rector\CodingStyle\Tests\Rector\FuncCall\VersionCompareFuncCallToConstantRector\Fixture;
 
 class VersionCompareClass
 {
@@ -20,7 +20,7 @@ class VersionCompareClass
 -----
 <?php
 
-namespace Rector\CodingStyle\Tests\Rector\FuncCall\FunctionCallToConstantRector\Fixture;
+namespace Rector\CodingStyle\Tests\Rector\FuncCall\VersionCompareFuncCallToConstantRector\Fixture;
 
 class VersionCompareClass
 {
